+++ test/adapter/web-standard/map-early-response.test.ts
@@ -359,7 +359,7 @@ describe('Web Standard - Map Early Response', () => {
 		expect(response).toBeInstanceOf(Response)
 		expect(response?.headers.get('accept-ranges')).toBeNull()
 		expect(response?.headers.get('content-range')).toBeNull()
-		expect(response?.status).toBe(200)
+		expect(response?.status).toBe(304)
 	})
 
 	it('map formdata', async () => {