+++ test/core/config.test.ts
@@ -14,4 +14,15 @@ describe('config', () => {
 
 		expect(response).toBe('a')
 	})
+
+	it('append prefix / if not provided', () => {
+		const plugin = new Elysia({ prefix: 'v1' }).get('thing', 'thing')
+
+		const app = new Elysia({ prefix: 'api' }).use(plugin)
+
+		expect(app.routes[0].path).toBe('/api/v1/thing')
+
+		// This should not error
+		app['~Routes']?.api.v1.thing
+	})
 })