+++ test/types/index.ts
@@ -2479,3 +2479,13 @@ type a = keyof {}
 			}
 		)
 }
+
+// append prefix / if not provided
+{
+	const plugin = new Elysia({ prefix: 'v1' }).get('thing', 'thing')
+
+	const app = new Elysia({ prefix: 'api' }).use(plugin)
+
+	// This should not error
+	app['~Routes']?.api.v1.thing
+}