+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Bug fix:
 - [#1316](https://github.com/elysiajs/elysia/pull/1316) fix context type when multiple macros are selected
 - [#1306](https://github.com/elysiajs/elysia/pull/1306) preserve type narrowing in getSchemaValidator
 - add `set` to `handleFile` when file is `ElysiaFile`
+- [Web Standard] inherit set.status for `ElysiaFile`
 
 # 1.3.7 - 31 Jul 2025
 Bug fix: