+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.3.9
+Bug fix:
+- [#459](https://github.com/elysiajs/elysia/issues/459) route prefix should give type error when prefix is not start with '/'
+
 # 1.3.8 - 31 Jul 2025
 Improvement:
 - ElysiaFile doesn't inherits `set.headers` eg. cors